requirements.txt
s3cmd
mcpi
python-dotenv
